+++ extension/src/process-editor/ivy-editor-provider.ts
@@ -4,7 +4,7 @@ import * as vscode from 'vscode';
 import fs from 'fs';
 import { NotificationType, RequestType, MessageParticipant } from 'vscode-messenger-common';
 import { messenger } from '../messenger';
-import { InscriptionWebSocketMessage, IvyScriptWebSocketMessage, WebSocketConnectionForwarder } from '../websocket-connection-forwarder';
+import { InscriptionWebSocketMessage, IvyScriptWebSocketMessage, WebSocketForwarder } from '../websocket-forwarder';
 
 const ColorThemeChangedNotification: NotificationType<'dark' | 'light'> = { method: 'colorThemeChanged' };
 const WebviewReadyNotification: NotificationType<void> = { method: 'ready' };
@@ -31,8 +31,8 @@ export default class IvyEditorProvider extends GlspEditorProvider {
 
     const messageParticipant = messenger.registerWebviewPanel(webviewPanel);
     const toDispose = new DisposableCollection(
-      new WebSocketConnectionForwarder(webSocketAddress + 'ivy-inscription-lsp', messageParticipant, InscriptionWebSocketMessage),
-      new WebSocketConnectionForwarder(webSocketAddress + 'ivy-script-lsp', messageParticipant, IvyScriptWebSocketMessage),
+      new WebSocketForwarder('ivy-inscription-lsp', messageParticipant, InscriptionWebSocketMessage),
+      new WebSocketForwarder('ivy-script-lsp', messageParticipant, IvyScriptWebSocketMessage),
       messenger.onNotification(WebviewReadyNotification, () => this.handleWebviewReadyNotification(webSocketAddress, messageParticipant), {
         sender: messageParticipant
       }),
@@ -65,7 +65,7 @@ export default class IvyEditorProvider extends GlspEditorProvider {
       `script-src 'nonce-${nonce}';` +
       `worker-src ${webview.cspSource};` +
       `font-src ${webview.cspSource};` +
-      `connect-src ${webview.cspSource} ${webSocketAddress}`;
+      `connect-src ${webview.cspSource}`;
 
     webviewPanel.webview.html = `
       <!DOCTYPE html>
